отредактировал написание запросов по правилам

+++ src/com/startjava/lesson_5/init_db.sql
@@ -4,16 +4,17 @@ CREATE DATABASE robots;
 
 DROP TABLE IF EXISTS jaegers;
 
-CREATE TABLE jaegers(
-	id SERIAL 	PRIMARY KEY,
-	model_name 	VARCHAR (20),
+CREATE TABLE jaegers (
+    id SERIAL 	PRIMARY KEY,
+	model_name 	VARCHAR(20),
 	mark		INT,
 	height		REAL,
-	weight		NUMERIC,
-	status		VARCHAR (16),
-	origin		VARCHAR (30),
+	weight		REAL,
+	status		VARCHAR(16),
+	origin		VARCHAR(30),
 	launch		DATE,
-	kaiju_kill	INT);
+	kaiju_kill	INT
+);
 	
 
 \ir 'populate.sql'
